Extracted SelectableChildrenList
